move node functionality to users route
