+++ src/controllers/user-controller.ts
@@ -5,13 +5,24 @@ import {
   UserDataSchema,
 } from "../schema/request-schemas/user-schema.js";
 import {
+  createNode,
   createUser,
+  deleteNode,
   getUserDataWithNodeTree,
   getUserFolders,
+  updateNodeName,
 } from "../db/user-queries.js";
 import bcrypt from "bcryptjs";
 import { setAuthCookie } from "../helpers/auth-helpers.js";
-import { GetUserFoldersSchema } from "../schema/request-schemas/node-schema.js";
+import {
+  CreateNodeSchema,
+  DeleteNodeSchema,
+  GetUserFoldersSchema,
+  UpdateNodeNameSchema,
+} from "../schema/request-schemas/node-schema.js";
+import { randomUUID } from "crypto";
+import cloudinary from "../config/cloudinary-config.js";
+import { CreateNode } from "../types/node-types.js";
 
 export const getUserDataWithNodeTreeGET = asyncHandler(async (req, res) => {
   schemaParser(UserDataSchema, req);
@@ -48,9 +59,57 @@ export const createUserPOST = asyncHandler(async (req, res, next) => {
   });
 });
 
-export const getUserFolderGET = asyncHandler(async (req, res) => {
+export const getUserFoldersGET = asyncHandler(async (req, res) => {
   schemaParser(GetUserFoldersSchema, req);
   const { userId } = req.params;
   const folders = await getUserFolders(+userId);
   res.json(folders);
+});
+
+export const uploadNodePOST = asyncHandler(async (req, res) => {
+  schemaParser(CreateNodeSchema, req);
+  //File is checked in schema
+  const file = req.file;
+  let fileLink: string | null = null;
+  let filePublicId: string | null = null;
+
+  if (file) {
+    const b64 = Buffer.from(file.buffer).toString("base64");
+    let dataURI = "data:" + file.mimetype + ";base64," + b64;
+
+    const fileName = file.originalname.replace(/\.[^.]*$/, "");
+    const result = await cloudinary.uploader.upload(dataURI, {
+      public_id: `${fileName}-${randomUUID()}`,
+      resource_type: "auto",
+      flags: "attachment",
+    });
+    fileLink = result.secure_url;
+    filePublicId = result.public_id;
+  }
+  const { parentNodeId, name, userId, type } = req.body;
+  const dataToSend: CreateNode = {
+    parentNodeId: +parentNodeId,
+    name,
+    userId: +userId,
+    type,
+    ...(!!fileLink && { fileLink }),
+    ...(!!filePublicId && { publicId: filePublicId }),
+  };
+  const node = await createNode(dataToSend);
+  res.json({ node });
+});
+
+export const updateNodeNamePUT = asyncHandler(async (req, res) => {
+  schemaParser(UpdateNodeNameSchema, req);
+  const { nodeId } = req.params;
+  const { newName } = req.body;
+  const node = await updateNodeName(+nodeId, newName);
+  res.json({ node });
+});
+
+export const deleteNodeDELETE = asyncHandler(async (req, res) => {
+  schemaParser(DeleteNodeSchema, req);
+  const { nodeId } = req.params;
+  await deleteNode(+nodeId);
+  res.json({ message: "Deleted successfully." });
 });
