+++ src/routes/node-router.ts
@@ -1,20 +1,13 @@
 import { Router } from "express";
 import isAuthenticated from "../middlewares/is-authenticated.js";
-import upload from "../config/multer-config.js";
 import {
   createSharedNodePOST,
-  deleteNodeDELETE,
   getSharedNodeGET,
-  updateNodeNamePUT,
-  uploadNodePOST,
 } from "../controllers/node-controller.js";
 
 const nodeRouter = Router();
 
 nodeRouter
-  .post("/", isAuthenticated, upload.single("file"), uploadNodePOST)
-  .put(":nodeId", isAuthenticated, updateNodeNamePUT)
-  .delete("/:nodeId", isAuthenticated, deleteNodeDELETE)
   .post("/shared", isAuthenticated, createSharedNodePOST)
   .get("/shared/:linkHash", getSharedNodeGET);
